Exit on errors

+++ mount-hard-drive.sh
@@ -36,8 +36,11 @@ done
 echo
 # sudo sfdisk --delete "${TARGET_DEVICE}" --backup
 sudo wipefs --all --force "${TARGET_DEVICE}"
+if [ $? -ne 0 ]; then exit
 echo
 echo "type=83" | sudo sfdisk "${TARGET_DEVICE}"  # create linux type partition
+if [ $? -ne 0 ]; then exit
 echo
 sudo mkfs.ext4 -F "${TARGET_PARTITION}"
+if [ $? -ne 0 ]; then exit
 echo
